+++ src/tm_devices/drivers/digital_multimeters/digital_multimeter.py
@@ -2,11 +2,11 @@
 
 from abc import ABC
 
-from tm_devices.driver_mixins.device_control.tsp_device import TSPDevice
 from tm_devices.helpers import DeviceTypes
 
 
-class DigitalMultimeter(TSPDevice, ABC):
+# pylint: disable=too-few-public-methods
+class DigitalMultimeter(ABC):  # noqa: B024
     """Base Digital Multimeter (DMM) device driver."""
 
     _DEVICE_TYPE = DeviceTypes.DMM.value