+++ src/tm_devices/drivers/source_measure_units/source_measure_unit.py
@@ -5,11 +5,11 @@ SMUs include TSP devices and PI devices.
 
 from abc import ABC
 
-from tm_devices.driver_mixins.device_control.tsp_device import TSPDevice
 from tm_devices.helpers import DeviceTypes
 
 
-class SourceMeasureUnit(TSPDevice, ABC):
+# pylint: disable=too-few-public-methods
+class SourceMeasureUnit(ABC):  # noqa: B024
     """Base Source Measure Unit (SMU) device driver."""
 
     _DEVICE_TYPE = DeviceTypes.SMU.value