+++ src/tm_devices/drivers/systems_switches/ss3706a.py
@@ -6,6 +6,7 @@ import pyvisa as visa
 
 from tm_devices.commands import SS3706AMixin
 from tm_devices.driver_mixins.device_control.device import family_base_class
+from tm_devices.driver_mixins.device_control.tsp_device import TSPDevice
 from tm_devices.drivers.systems_switches.systems_switch import SystemsSwitch
 from tm_devices.helpers import DeviceConfigEntry
 
@@ -14,7 +15,7 @@ from tm_devices.helpers import ReadOnlyCachedProperty as cached_property  # noqa
 
 
 @family_base_class
-class SS3706A(SS3706AMixin, SystemsSwitch):
+class SS3706A(SS3706AMixin, SystemsSwitch, TSPDevice):
     """SS3706A device driver."""
 
     ################################################################################################