+++ src/tm_devices/drivers/systems_switches/systems_switch.py
@@ -2,11 +2,11 @@
 
 from abc import ABC
 
-from tm_devices.driver_mixins.device_control.tsp_device import TSPDevice
 from tm_devices.helpers import DeviceTypes
 
 
-class SystemsSwitch(TSPDevice, ABC):
+# pylint: disable=too-few-public-methods
+class SystemsSwitch(ABC):  # noqa: B024
     """Base Systems Switch (SS) device driver."""
 
     _DEVICE_TYPE = DeviceTypes.SS.value