+++ tests/test_devices_legacy_tsp_ieee_cmds.py
@@ -1,11 +1,15 @@
 """Test the DMMs."""
 
+from typing import TYPE_CHECKING
 from unittest import mock
 
 import pytest
 import pyvisa as visa
 
 from tm_devices import DeviceManager
+
+if TYPE_CHECKING:
+    from tm_devices.drivers import DAQ6510, DMM6500, DMM7510
 
 
 def test_dmm6500(device_manager: DeviceManager) -> None:
@@ -14,7 +18,7 @@ def test_dmm6500(device_manager: DeviceManager) -> None:
     Args:
         device_manager: The DeviceManager object.
     """
-    dmm = device_manager.add_dmm("DMM6500-hostname", alias="dmm-device")
+    dmm: DMM6500 = device_manager.add_dmm("DMM6500-hostname", alias="dmm-device")
     assert id(device_manager.get_dmm(number_or_alias="dmm-device")) == id(dmm)
     assert id(device_manager.get_dmm(number_or_alias=dmm.device_number)) == id(dmm)
 
@@ -35,7 +39,7 @@ def test_dmm75xx(device_manager: DeviceManager) -> None:
     Args:
         device_manager: The DeviceManager object.
     """
-    dmm = device_manager.add_dmm("DMM7510-hostname")
+    dmm: DMM7510 = device_manager.add_dmm("DMM7510-hostname")
 
     with mock.patch("pyvisa.highlevel.VisaLibraryBase.clear", mock.MagicMock(return_value=None)):
         assert dmm.query_expect_timeout("INVALID?", timeout_ms=1) == ""
@@ -54,7 +58,7 @@ def test_daq6510(device_manager: DeviceManager) -> None:
     Args:
         device_manager: The DeviceManager object.
     """
-    daq = device_manager.add_daq("DAQ6510-hostname", alias="daq-device")
+    daq: DAQ6510 = device_manager.add_daq("DAQ6510-hostname", alias="daq-device")
     assert id(device_manager.get_daq(number_or_alias="daq-device")) == id(daq)
     assert id(device_manager.get_daq(number_or_alias=daq.device_number)) == id(daq)