+++ tests/test_ss.py
@@ -1,11 +1,15 @@
 """Test the SSs."""
 
+from typing import TYPE_CHECKING
 from unittest import mock
 
 import pytest
 import pyvisa as visa
 
 from tm_devices import DeviceManager
+
+if TYPE_CHECKING:
+    from tm_devices.drivers import SS3706A
 
 
 def test_ss(device_manager: DeviceManager) -> None:
@@ -14,7 +18,7 @@ def test_ss(device_manager: DeviceManager) -> None:
     Args:
         device_manager: The DeviceManager object.
     """
-    switch = device_manager.add_ss("ss3706a-hostname", alias="ss-device")
+    switch: SS3706A = device_manager.add_ss("ss3706a-hostname", alias="ss-device")
     assert id(device_manager.get_ss(number_or_alias="ss-device")) == id(switch)
     assert id(device_manager.get_ss(number_or_alias=switch.device_number)) == id(switch)
     assert switch.get_eventlog_status() == (True, '0,"No events to report - queue empty"')